requirements.txt
PyYAML~=6.0.1
pyTelegramBotAPI==4.17.0